requirements.txt
llama-cpp-python>=0.3.7
huggingface-hub>=0.28.1

